Removing pages in  and moving them down a level but using the basePath.
