+++ .env
@@ -2,5 +2,4 @@ NODE_ENV='production'
 PATRON_CREATION_URL=https://platform.nypl.org/api/v0.3/patrons
 PATRON_VALIDATION_URL=https://platform.nypl.org/api/v0.3/validations
 OAUTH_PROVIDER_URL=https://isso.nypl.org/oauth/token
-ASSET_PREFIX='card'
 PORT=3001
