+++ .env.example
@@ -7,4 +7,3 @@ OAUTH_CLIENT_SECRET=The secret this app uses to auth to the OAuth provider above
 NEXT_PUBLIC_USE_AXE='false' use 'true' if react-axe should be run while developing
 PORT=3001
 IPSTACK_KEY=key to api.ipstack service https://ipstack.com/
-ASSET_PREFIX='card'
