+++ src/logger/index.ts
@@ -1,8 +1,6 @@
-import winston from 'winston';
+import winston from "winston";
 
-const {
-  combine, timestamp, printf, colorize,
-} = winston.format;
+const { combine, timestamp, printf, colorize } = winston.format;
 const { File, Console } = winston.transports;
 
 // Set default NYPL agreed upon log levels
@@ -22,24 +20,24 @@ const nyplLogLevels = {
 
 const getLogLevelCode = (levelString) => {
   switch (levelString) {
-    case 'emergency':
+    case "emergency":
       return 0;
-    case 'alert':
+    case "alert":
       return 1;
-    case 'critical':
+    case "critical":
       return 2;
-    case 'error':
+    case "error":
       return 3;
-    case 'warning':
+    case "warning":
       return 4;
-    case 'notice':
+    case "notice":
       return 5;
-    case 'info':
+    case "info":
       return 6;
-    case 'debug':
+    case "debug":
       return 7;
     default:
-      return 'n/a';
+      return "n/a";
   }
 };
 
@@ -55,7 +53,7 @@ const nyplFormat = printf((options) => {
     level: options.level.toUpperCase(),
     message: options.message.toString(),
     // This is specific to this app to make searching easy.
-    appTag: 'library-card-app',
+    appTag: "library-card-app",
   };
 
   if (process.pid) {
@@ -71,7 +69,7 @@ const nyplFormat = printf((options) => {
 
 // The transport function that logs to a file.
 const fileTransport = new File({
-  filename: './log/library-card-app.log',
+  filename: "./log/library-card-app.log",
   handleExceptions: true,
   maxsize: 5242880, // 5MB
   maxFiles: 5,
@@ -85,13 +83,13 @@ const consoleTransport = new Console({
     nyplFormat,
     colorize({
       all: true,
-    }),
+    })
   ),
 });
 
 const loggerTransports: any[] = [consoleTransport];
 
-if (process.env.NODE_ENV !== 'test') {
+if (process.env.NODE_ENV !== "test") {
   loggerTransports.push(fileTransport);
 }
 
@@ -106,11 +104,11 @@ const logger = CreateLogger({
 });
 
 // Set the logger output level to one specified in the environment config.
-if (process.env.NODE_ENV === 'test') {
+if (process.env.NODE_ENV === "test") {
   // Console logs from tests can be disruptive.
-  logger.level = 'none';
+  logger.level = "none";
 } else {
-  logger.level = process.env.LOG_LEVEL || 'debug';
+  logger.level = process.env.LOG_LEVEL || "debug";
 }
 
 export default logger;
